+++ CHANGELOG.md
@@ -18,8 +18,16 @@ Security - in case of vulnerabilities.
 
 _TBD_
 
+### Added
+- [x] Added new properties to `Plan`: `pricing`, `recurringInterval`, `trial`, `setup`
+
 ### Changed
 - [x] Replaced magic methods like `get`, `post` with explicit defined methods.
+
+### Removed
+- [x] Removed deprecated Plan properties: `expiredTime`, `recurringAmount`, `trialAmount`, `setupAmount`,
+      `recurringPeriodUnit`, `recurringPeriodLength`, `recurringPeriodLimit`, `trialPeriodUnit`, `trialPeriodLength`,
+      `contractTermUnit`, `contractTermLength`, `minQuantity`, `maxQuantity` 
 
 ## [2.2.0] 2018-08-14